requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
rasa==3.6.0
pymongo==4.5.0
python-telegram-bot==20.5
cryptography==41.0.4
schedule==1.2.0
python-dotenv==1.0.0
requests==2.31.0
